refactor(tools): group internal utilities into support/ submodule
Move http_client.rs, parse.rs, and tolerant.rs into src/tools/support/
so they are clearly separated from the MCP tool modules alongside them.
No behavior change; only import paths updated.
