+++ src/tools/quote.rs
@@ -8,9 +8,9 @@ use rmcp::serde::Deserialize;
 
 use crate::counter::symbol_to_counter_id;
 use crate::error::Error;
-use crate::tools::http_client::http_get_tool;
-use crate::tools::parse;
-use crate::tools::tolerant::{
+use crate::tools::support::http_client::http_get_tool;
+use crate::tools::support::parse;
+use crate::tools::support::tolerant::{
     tolerant_bool, tolerant_i64, tolerant_option_usize, tolerant_option_vec_i32,
     tolerant_option_vec_string, tolerant_usize, tolerant_vec_string,
 };
