+++ src/tools/statement.rs
@@ -7,7 +7,7 @@ use rmcp::serde::Deserialize;
 use time::OffsetDateTime;
 
 use crate::error::Error;
-use crate::tools::tolerant::tolerant_option_i32;
+use crate::tools::support::tolerant::tolerant_option_i32;
 use crate::tools::tool_json;
 
 #[derive(Debug, Deserialize, JsonSchema)]
